% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.linreg}
\alias{predict.linreg}
\title{This contains the fitted values of linreg function.}
\usage{
\method{predict}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of linreg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the fitted values of linreg function.
}
