% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.linreg}
\alias{print.linreg}
\title{This contains the methods for linreg function.}
\usage{
\method{print}{linreg}(x, ...)
}
\description{
The available methods are :
print(), plot(), resid(), pred(), coef() and summary()
}
